Replace wallet lib, keep as legacy-wallet
