+++ client/legacy-wallet/assetregistry.go
@@ -1,4 +1,4 @@
-package alias_wallet
+package legacy_wallet
 
 import (
 	"github.com/iotaledger/hive.go/marshalutil"
