+++ client/legacy-wallet/connector.go
@@ -1,7 +1,7 @@
-package alias_wallet
+package legacy_wallet
 
 import (
-	"github.com/iotaledger/goshimmer/client/alias-wallet/packages/address"
+	"github.com/iotaledger/goshimmer/client/legacy-wallet/packages/address"
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/packages/mana"
 )
@@ -13,5 +13,4 @@ type Connector interface {
 	SendTransaction(transaction *ledgerstate.Transaction) (err error)
 	RequestFaucetFunds(address address.Address) (err error)
 	GetAllowedPledgeIDs() (pledgeIDMap map[mana.Type][]string, err error)
-	GetTransactionInclusionState(txID ledgerstate.TransactionID) (inc ledgerstate.InclusionState, err error)
 }
