+++ client/legacy-wallet/output.go
@@ -1,19 +1,36 @@
-package alias_wallet
+package legacy_wallet
 
 import (
-	"github.com/iotaledger/goshimmer/client/alias-wallet/packages/address"
-	"github.com/iotaledger/goshimmer/packages/ledgerstate"
-	"github.com/iotaledger/hive.go/stringify"
 	"time"
+
+	"github.com/iotaledger/hive.go/stringify"
+
+	"github.com/iotaledger/goshimmer/client/legacy-wallet/packages/address"
+	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 )
+
+// region Output ///////////////////////////////////////////////////////////////////////////////////////////////////////
 
 // Output is a wallet specific representation of an output in the IOTA network.
 type Output struct {
 	Address        address.Address
-	Object         ledgerstate.Output
+	OutputID       ledgerstate.OutputID
+	Balances       *ledgerstate.ColoredBalances
 	InclusionState InclusionState
 	Metadata       OutputMetadata
 }
+
+// String returns a human-readable representation of the Output.
+func (o *Output) String() string {
+	return stringify.Struct("Output",
+		stringify.StructField("Address", o.Address),
+		stringify.StructField("OutputID", o.OutputID),
+		stringify.StructField("Balances", o.Balances),
+		stringify.StructField("InclusionState", o.InclusionState),
+	)
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 // region InclusionState ///////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -68,11 +85,10 @@ func (o OutputsByID) OutputsByAddressAndOutputID() (outputsByAddressAndOutputID 
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-type OutputsByAddressAndOutputID map[address.Address]map[ledgerstate.OutputID]*Output
+// region OutputsByAddressAndOutputID //////////////////////////////////////////////////////////////////////////////////
 
-func NewAddressToOutputs() OutputsByAddressAndOutputID {
-	return make(map[address.Address]map[ledgerstate.OutputID]*Output)
-}
+// OutputsByAddressAndOutputID is a collection of Outputs associated with their Address and OutputID.
+type OutputsByAddressAndOutputID map[address.Address]map[ledgerstate.OutputID]*Output
 
 // OutputsByID returns a collection of Outputs associated with their OutputID.
 func (o OutputsByAddressAndOutputID) OutputsByID() (outputsByID OutputsByID) {
@@ -86,57 +102,4 @@ func (o OutputsByAddressAndOutputID) OutputsByID() (outputsByID OutputsByID) {
 	return
 }
 
-func (o OutputsByAddressAndOutputID) ValueOutputsOnly() OutputsByAddressAndOutputID {
-	result := NewAddressToOutputs()
-	for addy, IDToOutputMap := range o {
-		for outputID, output := range IDToOutputMap {
-			switch output.Object.Type() {
-			case ledgerstate.SigLockedSingleOutputType, ledgerstate.SigLockedColoredOutputType, ledgerstate.ExtendedLockedOutputType:
-				if _, addressExists := result[addy]; !addressExists {
-					result[addy] = make(map[ledgerstate.OutputID]*Output)
-				}
-				result[addy][outputID] = output
-			}
-		}
-	}
-	return result
-}
-
-func (o OutputsByAddressAndOutputID) AliasOutputsOnly() OutputsByAddressAndOutputID {
-	result := NewAddressToOutputs()
-	for addy, IDToOutputMap := range o {
-		for outputID, output := range IDToOutputMap {
-			switch output.Object.Type() {
-			case ledgerstate.AliasOutputType:
-				if _, addressExists := result[addy]; !addressExists {
-					result[addy] = make(map[ledgerstate.OutputID]*Output)
-				}
-				result[addy][outputID] = output
-			}
-		}
-	}
-	return result
-}
-
-func (o OutputsByAddressAndOutputID) TotalFundsInOutputs() map[ledgerstate.Color]uint64 {
-	result := make(map[ledgerstate.Color]uint64)
-	for _, IDToOutputMap := range o {
-		for _, output := range IDToOutputMap {
-			output.Object.Balances().ForEach(func(color ledgerstate.Color, balance uint64) bool {
-				result[color] += balance
-				return true
-			})
-		}
-	}
-	return result
-}
-
-func (o OutputsByAddressAndOutputID) ToLedgerStateOutputs() ledgerstate.Outputs {
-	outputs := ledgerstate.Outputs{}
-	for _, outputIDMapping := range o {
-		for _, output := range outputIDMapping {
-			outputs = append(outputs, output.Object)
-		}
-	}
-	return outputs
-}
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
