+++ client/legacy-wallet/packages/seed/seed.go
@@ -1,9 +1,9 @@
 package seed
 
 import (
-	address "github.com/iotaledger/goshimmer/client/alias-wallet/packages/address"
 	"github.com/iotaledger/hive.go/crypto/ed25519"
 
+	"github.com/iotaledger/goshimmer/client/legacy-wallet/packages/address"
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 )
