+++ client/wallet/packages/delegatefunds_options/delegatefunds_options.go
@@ -5,7 +5,7 @@ import (
 	"golang.org/x/xerrors"
 	"time"
 
-	"github.com/iotaledger/goshimmer/client/alias-wallet/packages/address"
+	"github.com/iotaledger/goshimmer/client/wallet/packages/address"
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 )
