+++ tools/cli-wallet/create_nft.go
@@ -1,0 +1,1 @@
+package main
